+++ src/WindowsDeviceManagerAgent/DatabaseWriter.cs
@@ -18,7 +18,7 @@ namespace WindowsDeviceManagerAgent
         /// <summary>
         /// データベースuser_version
         /// </summary>
-        private static readonly string _databaseUserVersion = "4";
+        private static readonly string _databaseUserVersion = "5";
 
         /// <summary>
         /// SQLコマンド(テーブル作成)
@@ -35,14 +35,15 @@ namespace WindowsDeviceManagerAgent
                                                              + "BIOSManufacturer TEXT, "
                                                              + "BIOSVersion TEXT, "
                                                              + "BitLockerStatus TEXT, "
+                                                             + "AntiVirusSoftware TEXT, "
                                                              + "LastUpdate TEXT)";
 
         /// <summary>
         /// SQLコマンド(レコード登録)
         /// </summary>
         private static readonly string _insertCommand = "INSERT OR REPLACE INTO WindowsDeviceInfo"
-                                                        + "(HostName, UserName, OSName, OSBuildNumber, OSVersion, ComputerManufacturer, ComputerModel, Processor, BIOSManufacturer, BIOSVersion, BitLockerStatus, LastUpdate) "
-                                                        + "VALUES(@p_HostName, @p_UserName, @p_OSName, @p_OSBuildNumber, @p_OSVersion, @p_ComputerManufacturer, @p_ComputerModel, @p_Processor, @p_BIOSManufacturer, @p_BIOSVersion, @p_BitLockerStatus, @p_LastUpdate)";
+                                                        + "(HostName, UserName, OSName, OSBuildNumber, OSVersion, ComputerManufacturer, ComputerModel, Processor, BIOSManufacturer, BIOSVersion, BitLockerStatus, AntiVirusSoftware, LastUpdate) "
+                                                        + "VALUES(@p_HostName, @p_UserName, @p_OSName, @p_OSBuildNumber, @p_OSVersion, @p_ComputerManufacturer, @p_ComputerModel, @p_Processor, @p_BIOSManufacturer, @p_BIOSVersion, @p_BitLockerStatus, @p_AntiVirusSoftware, @p_LastUpdate)";
 
         //--------------------------------------------------
         // メソッド
@@ -81,6 +82,7 @@ namespace WindowsDeviceManagerAgent
                 _ = command.Parameters.Add(new SQLiteParameter("@p_BIOSManufacturer", writeValue.BIOSManufacturer));
                 _ = command.Parameters.Add(new SQLiteParameter("@p_BIOSVersion", writeValue.BIOSVersion));
                 _ = command.Parameters.Add(new SQLiteParameter("@p_BitLockerStatus", writeValue.BitLockerStatus));
+                _ = command.Parameters.Add(new SQLiteParameter("@p_AntiVirusSoftware", writeValue.AntiVirusSoftware));
                 _ = command.Parameters.Add(new SQLiteParameter("@p_LastUpdate", writeValue.LastUpdate));
                 command.Prepare();
                 _ = command.ExecuteNonQuery();
@@ -101,37 +103,47 @@ namespace WindowsDeviceManagerAgent
                 switch (version)
                 {
                     case "0":
-                        // user_versionが0のときは､0から4に更新する
+                        // user_versionが0のときは､0から5に更新する
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageDetectOldDatabase);
                         UpdateDatabaseVersion1();
                         UpdateDatabaseVersion2();
                         UpdateDatabaseVersion3();
                         UpdateDatabaseVersion4();
+                        UpdateDatabaseVersion5();
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageUpdateDatabaseComplete);
                         break;
                     case "1":
-                        // user_versionが1のときは､1から4に更新する
+                        // user_versionが1のときは､1から5に更新する
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageDetectOldDatabase);
                         UpdateDatabaseVersion2();
                         UpdateDatabaseVersion3();
                         UpdateDatabaseVersion4();
+                        UpdateDatabaseVersion5();
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageUpdateDatabaseComplete);
                         break;
                     case "2":
-                        // user_versionが2のときは､2から4に更新する
+                        // user_versionが2のときは､2から5に更新する
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageDetectOldDatabase);
                         UpdateDatabaseVersion3();
                         UpdateDatabaseVersion4();
+                        UpdateDatabaseVersion5();
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageUpdateDatabaseComplete);
                         break;
                     case "3":
-                        // user_versionが3のときは､3から4に更新する
+                        // user_versionが3のときは､3から5に更新する
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageDetectOldDatabase);
                         UpdateDatabaseVersion4();
+                        UpdateDatabaseVersion5();
                         ConsoleWrapper.WriteLine(Resources.Strings.MessageUpdateDatabaseComplete);
                         break;
                     case "4":
-                        // user_versionが4のときは最新のため更新不要
+                        // user_versionが4のときは､4から5に更新する
+                        ConsoleWrapper.WriteLine(Resources.Strings.MessageDetectOldDatabase);
+                        UpdateDatabaseVersion5();
+                        ConsoleWrapper.WriteLine(Resources.Strings.MessageUpdateDatabaseComplete);
+                        break;
+                    case "5":
+                        // user_versionが5のときは最新のため更新不要
                         break;
                     default:
                         // user_versionが想定外のときは更新不要
@@ -252,6 +264,23 @@ namespace WindowsDeviceManagerAgent
                 command.CommandText = "ALTER TABLE WindowsDeviceInfo ADD COLUMN BitLockerStatus TEXT";
                 _ = command.ExecuteNonQuery();
                 command.CommandText = $"PRAGMA user_version = 4";
+                _ = command.ExecuteNonQuery();
+            }
+            connection.Close();
+        }
+
+        /// <summary>
+        /// データベースファイル更新処理(user_version5化)
+        /// </summary>
+        private static void UpdateDatabaseVersion5()
+        {
+            using SQLiteConnection connection = new($"Data Source = {_databaseFileName}");
+            connection.Open();
+            using (SQLiteCommand command = connection.CreateCommand())
+            {
+                command.CommandText = "ALTER TABLE WindowsDeviceInfo ADD COLUMN AntiVirusSoftware TEXT";
+                _ = command.ExecuteNonQuery();
+                command.CommandText = $"PRAGMA user_version = 5";
                 _ = command.ExecuteNonQuery();
             }
             connection.Close();