+++ src/WindowsDeviceManagerViewer/Resources/Strings.Designer.cs
@@ -61,6 +61,15 @@ namespace WindowsDeviceManagerViewer.Resources {
         }
         
         /// <summary>
+        ///   アンチウィルスソフトウェア に類似しているローカライズされた文字列を検索します。
+        /// </summary>
+        public static string AntiVirusSoftware {
+            get {
+                return ResourceManager.GetString("AntiVirusSoftware", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   WindowsDeviceManagerViewer に類似しているローカライズされた文字列を検索します。
         /// </summary>
         public static string ApplicationName {